requirements.txt
Flask
gunicorn
firebase-admin
python-dotenv
Flask-Cors

